requirements.txt
neo4j==5.28.1
requests==2.32.3
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
streamlit==1.45.0
